% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data_wrangling.R
\name{extract_comm_from_bdforet}
\alias{extract_comm_from_bdforet}
\title{Extract Commune Area from BDForêt Dataset}
\usage{
extract_comm_from_bdforet(bd_path, comm_geom)
}
\arguments{
\item{bd_path}{Character. Path to the BDForêt shapefile.}

\item{comm_geom}{sf object. Geometry of the commune in WGS84 CRS.}
}
\value{
sf object. Extracted geometry of the commune area with cleaned polygons.
}
\description{
This function extracts the specified commune area from the BDForêt dataset by intersecting
the commune geometry with the provided forest shapefile.
}
\examples{
\dontrun{
commune_area <- extract_comm_from_bdforet("path/to/bdforet.shp", commune_geom)
}
}
