% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_utils.R
\name{get_bounding_box}
\alias{get_bounding_box}
\title{Get bounding box around a point}
\usage{
get_bounding_box(lat, lon, radius_km = 1)
}
\arguments{
\item{lat}{}

\item{lon}{}

\item{radius_km}{}
}
\description{
Get bounding box around a point
}
