% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{get_leaflet_map}
\alias{get_leaflet_map}
\title{Generate Leaflet Map with Change Regions and Forest Types}
\usage{
get_leaflet_map(cr_mask, bd_shp_cropped, comm_geom)
}
\arguments{
\item{cr_mask}{sf object. Spatial polygons of change regions with attributes.}

\item{bd_shp_cropped}{sf object. Cropped shapefile of forest areas with BDForêt attributes.}

\item{comm_geom}{sf object. Geometry of the commune to be outlined.}
}
\value{
leaflet object. Interactive map with layers for change regions, forest types, and commune boundaries.
}
\description{
This function creates an interactive Leaflet map displaying change regions with NDVI thresholding,
along with forest type areas from BDForêt data.
}
\examples{
\dontrun{
map <- get_leaflet_map(change_regions, forest_types, commune_boundary)
}
}
