% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{get_ndvi_diff}
\alias{get_ndvi_diff}
\title{Calculate NDVI Differences Between Years}
\usage{
get_ndvi_diff(
  ndvi.y,
  n_years_lag = 1,
  max_difference_days = 15,
  min_past_ndvi = 0.7
)
}
\arguments{
\item{ndvi.y}{terra::SpatRaster. Aggregated NDVI data with monthly time steps.}

\item{n_years_lag}{Integer. Lag period in years for comparison. Default is 1.}

\item{max_difference_days}{Integer. Maximum allowable days for difference matching. Default is 15.}

\item{min_past_ndvi}{Numeric. Minimum NDVI threshold for historical comparison. Default is 0.7.}
}
\value{
terra::SpatRaster. NDVI difference layers.
}
\description{
Calculates the differences in NDVI values over a specified lag period, filtering based on a minimum NDVI threshold.
}
\examples{
\dontrun{
ndvi_diff <- get_ndvi_diff(ndvi_agg, n_years_lag = 1, min_past_ndvi = 0.7)
}
}
