% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_utils.R
\name{get_token}
\alias{get_token}
\title{Retrieve Sentinel Hub Authentication Token}
\usage{
get_token(client_id, client_secret)
}
\arguments{
\item{client_id}{Character. Sentinel Hub client ID.}

\item{client_secret}{Character. Sentinel Hub client secret.}
}
\value{
Character. Access token required for API calls.
}
\description{
This function retrieves an authentication token for accessing Sentinel Hub API.
}
\examples{
\dontrun{
token <- get_token("your_client_id", "your_client_secret")
}
}
