% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_data_wrangling.R
\name{transform_bbox}
\alias{transform_bbox}
\title{Transform and Adjust Bounding Box}
\usage{
transform_bbox(bbox, crs_from = 4326, crs_to = 32736, rounding = -1)
}
\arguments{
\item{bbox}{Numeric vector. Bounding box coordinates as \code{c(xmin, ymin, xmax, ymax)}.}

\item{crs_from}{Integer. EPSG code of the source CRS. Default is 4326 (WGS84).}

\item{crs_to}{Integer. EPSG code of the target CRS. Default is 32736 (UTM zone 36S).}

\item{rounding}{Integer. Rounding precision for the transformed bounding box coordinates. Default is -1.}
}
\value{
List. Contains the transformed bounding box as an sf object, and numeric width and height of the transformed box.
}
\description{
Transforms the bounding box to a specified CRS, calculates width and height, and rounds coordinates.
}
\examples{
\dontrun{
transformed_bbox <- transform_bbox(c(2.0, 48.0, 3.0, 49.0), crs_from = 4326, crs_to = 32631)
}
}
