% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel_utils.R
\name{write_extract_tar_response}
\alias{write_extract_tar_response}
\title{Write and Extract TAR File from API Response}
\usage{
write_extract_tar_response(response, output_dir = ".")
}
\arguments{
\item{response}{response object. API response containing the TAR file data.}

\item{output_dir}{Character. Directory path to save and extract the contents of the TAR file. Default is current directory.}
}
\value{
Character vector. Names of the extracted files.
}
\description{
This function writes a TAR file from the raw content of an API response and extracts
the contents to the specified output directory.
}
\examples{
\dontrun{
file_names <- write_extract_tar_response(response, "output_directory")
}
}
